splitting pyxis calls
