+++ cmd/cve-analyser/main.go
@@ -5,16 +5,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"strings"
 	"sync"
 
-	hydra_api "github.com/p-rog/cve-analyser/pkg"
+	helper "github.com/p-rog/cve-analyser/pkg"
 )
-
-const PyxisUrl string = "https://catalog.redhat.com/api/containers/v1"
 
 type Output struct {
 	CVE         string
@@ -23,63 +19,9 @@ type Output struct {
 	Information string
 }
 
-type PyxisResponse struct {
-	Data []struct {
-		Arch string   `json:"architecture"`
-		CPEs []string `json:"cpe_ids"`
-	} `json:"data"`
-}
-
 func PrettyPrint(i interface{}) string {
 	s, _ := json.MarshalIndent(i, "", "\t")
 	return string(s)
-}
-
-func unique(s []string) []string {
-	inResult := make(map[string]bool)
-	var result []string
-	for _, str := range s {
-		if _, ok := inResult[str]; !ok {
-			inResult[str] = true
-			result = append(result, str)
-		}
-	}
-	return result
-}
-
-func CallPyxis(repository string, name string, tag string) []string {
-	url := PyxisUrl + "/repositories/registry/registry.access.redhat.com/repository/" + repository + "/" + name + "/tag/" + tag
-	//fmt.Println(url)
-	// Get request
-	resp, err := http.Get(url)
-	if err != nil {
-		fmt.Println("No response from request")
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body) // response body is []byte
-
-	var result PyxisResponse
-	if err := json.Unmarshal(body, &result); err != nil { // Parse []byte to the go struct pointer
-		fmt.Println("Can not unmarshal JSON")
-	}
-
-	var dirty_cpe_ids []string
-	for _, look := range result.Data {
-		if look.Arch == "amd64" {
-			dirty_cpe_ids = append(dirty_cpe_ids, look.CPEs...)
-		}
-	}
-
-	// generate list of CPEs with sw_edition and without
-	var cpe_ids []string
-	var clean_cpe_ids []string
-	for _, look := range dirty_cpe_ids {
-		clean_cpe_ids = strings.Split(look, "::")
-		cpe_ids = append(cpe_ids, clean_cpe_ids[0])
-		cpe_ids = append(cpe_ids, look)
-	}
-
-	return unique(cpe_ids)
 }
 
 func fileExist(file_name string) bool {
@@ -96,7 +38,7 @@ func processLines(wg *sync.WaitGroup, jobs <-chan []string, results chan<- Outpu
 		//Collecting and grepping security data
 		cve_id := line_data[0]
 		//fmt.Printf("%v", cve_id)
-		result := hydra_api.CallHydra(cve_id)
+		result := helper.CallHydra(cve_id)
 
 		var image_Repo string
 		var image_Name string
@@ -121,7 +63,7 @@ func processLines(wg *sync.WaitGroup, jobs <-chan []string, results chan<- Outpu
 			}
 			//fmt.Printf("%v\n%v\n%v\n", image_Repo, image_Name, image_Tag)
 
-			image_cpe_info = CallPyxis(image_Repo, image_Name, image_Tag)
+			image_cpe_info = helper.CallPyxis_CPE(image_Repo, image_Name, image_Tag)
 			//fmt.Println(PrettyPrint(image_cpe_info))
 
 			//Searching for non_rpm fixes based on the container name
