+++ pkg/hydra_api.go
@@ -1,4 +1,4 @@
-package hydra_api
+package helper
 
 import (
 	"encoding/json"
